# baggage: more efficient member validation