Instead of performing many, many comparisons for each character in a key and value, create an array of 128 bits and test for membership in that array, which is quicker.

Benchmarks on a Mac M1, ran 20 times.

```
name                             old time/op    new time/op    delta
New-8                               747ns ± 4%     741ns ± 3%    ~     (p=0.089 n=17+19)
NewMemberRaw-8                     11.4ns ± 1%    10.9ns ± 1%  -4.67%  (p=0.000 n=17+19)
Parse-8                             619ns ± 2%     603ns ± 1%  -2.54%  (p=0.000 n=18+18)
String-8                           1.12µs ± 4%    1.09µs ± 2%  -2.65%  (p=0.000 n=19+18)
ValueEscape/nothing_to_escape-8    6.93ns ± 1%    6.94ns ± 0%  +0.18%  (p=0.012 n=18+20)
ValueEscape/requires_escaping-8    26.3ns ± 2%    26.1ns ± 1%  -0.91%  (p=0.000 n=19+19)
ValueEscape/long_value-8            641ns ± 4%     595ns ± 1%  -7.14%  (p=0.000 n=18+17)
```

The approach is taken from the same approach in `encoding/json` in the standard library, which I also contributed.