@pellared 2024-04-16
> I don't think this PR needs a changelog entry.

We tend to add changelog entries when we improve performance.