@pellared 2024-04-16
> The approach is taken from the same approach in encoding/json in the standard library, which I also contributed.

Can you please provide a reference?