@pellared 2024-04-18
@open-telemetry/go-approvers waiting 1 day before merging.